[FEAT] 주문요청시 주문 상태 요청됨으로 변경
