+++ src/main/java/com/devloop/order/repository/OrderRepository.java
@@ -3,5 +3,8 @@ package com.devloop.order.repository;
 import com.devloop.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface OrderRepository extends JpaRepository<Order, Long> {
+    Optional<Order> findByOrderRequestId(String orderRequestId);
 }
