+++ src/main/resources/templates/payment-request.html
@@ -36,7 +36,7 @@
         const customerEmail = `[[${customerEmail}]]`
         const amount = `[[${totalPrice}]]`
         const orderRequestId = `[[${orderRequestId}]]`
-        const customerKey = "qalPSYonw4F919YCypHmk";
+        const customerKey = `[[${customerKey}]]`;
 
         const widgets = tossPayments.widgets({
             customerKey,
@@ -63,11 +63,10 @@
             await widgets.requestPayment({
                 orderId: orderRequestId,
                 orderName: orderName,
-                successUrl: window.location.origin + "/payments-success",
-                failUrl: window.location.origin + "/fail",
+                successUrl: window.location.origin + "/payments-success?orderRequestId="+orderRequestId,
+                failUrl: window.location.origin + "/api/v2/orders-fail?orderRequestId="+orderRequestId,
                 customerEmail: customerEmail,
                 customerName: customerName,
-                customerMobilePhone: "01012341234",
             });
         });
     }
